@@ -1,28 +1,28 @@
 module github.com/insomniacslk/dhcp
 
-go 1.18
+go 1.20
 
 require (
 	github.com/fanliao/go-promise v0.0.0-20141029170127-1890db352a72
 	github.com/google/go-cmp v0.5.9
 	github.com/hugelgupf/socketpair v0.0.0-20190730060125-05d35a94e714
-	github.com/jsimonetti/rtnetlink v0.0.0-20201110080708-d2c240429e6c
-	github.com/mdlayher/netlink v1.1.1
-	github.com/mdlayher/packet v1.1.1
+	github.com/jsimonetti/rtnetlink v1.3.5
+	github.com/mdlayher/netlink v1.7.2
+	github.com/mdlayher/packet v1.1.2
 	github.com/stretchr/testify v1.6.1
 	github.com/u-root/uio v0.0.0-20230220225925-ffce2a382923
-	golang.org/x/net v0.7.0
-	golang.org/x/sys v0.5.0
+	golang.org/x/net v0.17.0
+	golang.org/x/sys v0.13.0
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.0 // indirect
 	github.com/josharian/native v1.1.0 // indirect
-	github.com/mdlayher/socket v0.4.0 // indirect
+	github.com/mdlayher/socket v0.4.1 // indirect
 	github.com/pierrec/lz4/v4 v4.1.14 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/smartystreets/goconvey v1.6.4 // indirect
 	github.com/stretchr/objx v0.1.0 // indirect
-	golang.org/x/sync v0.1.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	golang.org/x/sync v0.3.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0 // indirect
 )
