@@ -3,7 +3,6 @@ package netboot
 import (
 	"errors"
 	"fmt"
-	"io/ioutil"
 	"net"
 	"os"
 	"strings"
@@ -211,7 +210,7 @@ func ConfigureInterface(ifname string, netconf *NetConf) (err error) {
 	if len(netconf.DNSSearchList) > 0 {
 		resolvconf += fmt.Sprintf("search %s\n", strings.Join(netconf.DNSSearchList, " "))
 	}
-	if err = ioutil.WriteFile("/etc/resolv.conf", []byte(resolvconf), 0644); err != nil {
+	if err = os.WriteFile("/etc/resolv.conf", []byte(resolvconf), 0644); err != nil {
 		return fmt.Errorf("could not write resolv.conf file %v", err)
 	}
 
